+++ ezomero/__init__.py
@@ -14,9 +14,6 @@ from ._ezomero import (post_dataset,
                        get_user_id,
                        get_original_filepaths,
                        put_map_annotation,
-                       filter_by_filename,
-                       link_images_to_dataset,
-                       link_datasets_to_project,
                        print_map_annotation,
                        print_groups,
                        print_projects,
@@ -24,6 +21,9 @@ from ._ezomero import (post_dataset,
                        connect,
                        store_connection_params,
                        set_group)
+from ._misc import (filter_by_filename,
+                    link_images_to_dataset,
+                    link_datasets_to_project)
 
 __all__ = ['post_dataset',
            'post_image',
