+++ ezomero/_ezomero.py
@@ -9,8 +9,7 @@ from getpass import getpass
 from omero.gateway import BlitzGateway
 from omero.gateway import MapAnnotationWrapper, FileAnnotationWrapper
 from omero.gateway import ProjectWrapper, DatasetWrapper
-from omero.model import MapAnnotationI, DatasetI, ProjectI, ProjectDatasetLinkI
-from omero.model import DatasetImageLinkI, ImageI, ExperimenterI
+from omero.model import MapAnnotationI, DatasetI, ProjectI
 from omero.model import RoiI, PointI, LineI, RectangleI, EllipseI
 from omero.model import PolygonI, LengthI, enums
 from omero.rtypes import rlong, rstring, rint, rdouble
@@ -18,6 +17,7 @@ from omero.sys import Parameters
 from omero import ApiUsageException
 from ezomero.rois import Point, Line, Rectangle, Ellipse, Polygon
 from pathlib import Path
+from ._misc import link_datasets_to_project
 
 
 def get_default_args(func):
@@ -1176,108 +1176,6 @@ def put_map_annotation(conn, map_ann_id, kv_dict, ns=None, across_groups=True):
     return None
 
 
-# filters
-def filter_by_filename(conn, im_ids, imported_filename):
-    """Filter list of image ids by originalFile name
-
-    Sometimes we know the filename of an image that has been imported into
-    OMERO but not necessarily the image ID. This is frequently the case when
-    we want to annotate a recently imported image. This funciton will help
-    to filter a list of image IDs to only those associated with a particular
-    filename.
-
-    Parameters
-    ----------
-    conn : ``omero.gateway.BlitzGateway`` object
-        OMERO connection.
-    im_ids : list of int
-        List of OMERO image IDs.
-    imported_filename : str
-        The full filename (with extension) of the file whose OMERO image
-        we are looking for. NOT the path of the image.
-
-    Returns
-    -------
-    filtered_im_ids : list of int
-        Filtered list of images with originalFile name matching
-        ``imported_filename``.
-
-    Notes
-    -----
-    This function should be used as a filter on an image list that has been
-    already narrowed down as much as possible. Note that many different images
-    in OMERO may share the same filename (e.g., image.tif).
-
-    Examples
-    --------
-    >>> im_ids = get_image_ids(conn, dataset=303)
-    >>> im_ids = filter_by_filename(conn, im_ids, "feb_2020.tif")]
-    """
-
-    q = conn.getQueryService()
-    params = Parameters()
-    params.map = {"oname": rstring(imported_filename)}
-    results = q.projection(
-        "SELECT i.id FROM Image i"
-        " JOIN i.fileset fs"
-        " JOIN fs.usedFiles u"
-        " JOIN u.originalFile o"
-        " WHERE o.name=:oname",
-        params,
-        conn.SERVICE_OPTS
-        )
-    im_id_matches = [r[0].val for r in results]
-
-    return list(set(im_ids) & set(im_id_matches))
-
-
-# linking functions
-def link_images_to_dataset(conn, image_ids, dataset_id):
-    """Link images to the specified dataset.
-
-    Nothing is returned by this function.
-
-    Parameters
-    ----------
-    conn : ``omero.gateway.BlitzGateway`` object
-        OMERO connection.
-    im_ids : list of int
-        List of OMERO image IDs.
-    dataset_id : int
-        Id of dataset to which images will be linked.
-    """
-    user_id = _get_current_user(conn)
-    for im_id in image_ids:
-        link = DatasetImageLinkI()
-        link.setParent(DatasetI(dataset_id, False))
-        link.setChild(ImageI(im_id, False))
-        link.details.owner = ExperimenterI(user_id, False)
-        conn.getUpdateService().saveObject(link, conn.SERVICE_OPTS)
-
-
-def link_datasets_to_project(conn, dataset_ids, project_id):
-    """Link datasets to the specified project.
-
-    Nothing is returned by this function.
-
-    Parameters
-    ----------
-    conn : ``omero.gateway.BlitzGateway`` object
-        OMERO connection.
-    im_ids : list of int
-        List of OMERO Dataset Ids.
-    dataset_id : int
-        Id of Project to which Datasets will be linked.
-    """
-    user_id = _get_current_user(conn)
-    for did in dataset_ids:
-        link = ProjectDatasetLinkI()
-        link.setParent(ProjectI(project_id, False))
-        link.setChild(DatasetI(did, False))
-        link.details.owner = ExperimenterI(user_id, False)
-        conn.getUpdateService().saveObject(link, conn.SERVICE_OPTS)
-
-
 # prints
 def print_map_annotation(conn, map_ann_id):
     """Print some information and value of a map annotation.
@@ -1617,11 +1515,3 @@ def set_group(conn, group_id):
     else:
         logging.warning(f'User {user_id} is not a member of Group {group_id}')
         return False
-
-
-# private functions
-def _get_current_user(conn):
-    userid = conn.SERVICE_OPTS.getOmeroUser()
-    if userid is None:
-        userid = conn.getUserId()
-    return userid
